+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.3 - 2021-04-24
+## Add
+- Psalm as Static Code Analysis Tool
+- Add composer install --no-dev before deploy
+
 ## 0.3.2 - 2021-04-22
 ### Add
 - Postgresql template option